requirements.txt
openai==0.1.2
python-dotenv==0.19.2
requests==2.27.1